requirements.txt
altair==6.0.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.1
attrs==26.1.0
backports.zstd==1.3.0
bertopic==0.17.4
blinker==1.9.0
Brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
filetype==1.2.0
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
gmpy2==2.3.0
h2==4.3.0
hdbscan==0.8.41
hf-xet==1.4.2
hpack==4.1.0
huggingface_hub==0.36.0
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
importlib_metadata==8.8.0
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.2.1
minify_html==0.18.1
mpmath==1.3.0
multimethod==1.12
narwhals==2.18.0
nest_asyncio==1.6.0
networkx==3.6.1
numba==0.62.1
numpy==2.3.5
optree==0.19.0
packaging==25.0
pandas==2.3.3
parso==0.8.6
patsy==1.0.2
pexpect==4.9.0
phik==0.12.5
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==23.0.1
pybind11==3.0.1
pybind11-global==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pynndescent==0.5.13
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyWavelets==1.9.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.3.0
setuptools==80.10.2
six==1.17.0
smmap==5.0.2
stack_data==0.6.3
statsmodels==0.14.6
streamlit==1.55.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typeguard==4.5.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
umap-learn==0.5.11
urllib3==2.6.3
visions==0.8.1
watchdog==6.0.0
wcwidth==0.6.0
wheel==0.46.3
wordcloud==1.9.6
ydata-profiling==4.18.1
zipp==3.23.0
